Separate initial conditions routines into separate file.
